requirements.txt
anyio==4.9.0
apsw==3.49.1.0
apswutils==0.0.2
beautifulsoup4==4.13.3
boto3==1.37.13
botocore==1.37.13
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
cssutils==2.11.1
fastcore==1.7.29
fastlite==0.1.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jmespath==1.0.1
lucide-fasthtml==0.0.9
lxml==5.3.1
more-itertools==10.6.0
oauthlib==3.2.2
packaging==24.2
premailer==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.12.4
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
yagmail==0.15.293
